fix: 連続時限のIF項目名修正

+++ src/domain/vo/constraint_definition.py
@@ -5,5 +5,5 @@ from pydantic import BaseModel
 class ConstraintDefinitionVo(BaseModel):
     code: str
     soft_flag: bool
-    penalty_weight: Optional[int]
+    penalty_weight: Optional[float]
     parameters: Optional[Dict[str, str]]
